refactor: Group Dto 폴더 이동
