+++ src/main/java/com/dnd/moddo/domain/group/service/implementation/GroupCreater.java
@@ -1,7 +1,7 @@
 package com.dnd.moddo.domain.group.service.implementation;
 
-import com.dnd.moddo.domain.group.dto.GroupRequest;
-import com.dnd.moddo.domain.group.dto.GroupResponse;
+import com.dnd.moddo.domain.group.dto.request.GroupRequest;
+import com.dnd.moddo.domain.group.dto.response.GroupResponse;
 import com.dnd.moddo.domain.group.entity.Group;
 import com.dnd.moddo.domain.group.repository.GroupRepository;
 import com.dnd.moddo.domain.user.entity.User;
